[FLINK-1159] Refactoring to include all acceptPartialFunctions implementation with a single import
