+++ flink-scala/src/main/scala/org/apache/flink/api/scala/extensions/impl/acceptPartialFunctions/OnCrossDataSet.scala
@@ -15,27 +15,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.api.scala.extensions.acceptPartialFunctions
+package org.apache.flink.api.scala.extensions.impl.acceptPartialFunctions
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.api.scala.{DataSet, CoGroupDataSet}
+import org.apache.flink.api.scala.{CrossDataSet, DataSet}
 
 import scala.reflect.ClassTag
 
-class OnCoGroupDataSet[L: TypeInformation, R: TypeInformation](dataset: CoGroupDataSet[L, R]) {
+class OnCrossDataSet[L: TypeInformation, R: TypeInformation](ds: CrossDataSet[L, R]) {
 
   /**
-    * Co-groups the data sets using the function `fun` to project elements from both in
-    * the resulting data set
+    * Starting from a cross data set, uses the function `fun` to project elements from
+    * both the input data sets in the resulting data set
     *
-    * @param fun The function that defines the projection of the co-group operation
+    * @param fun The function that defines the projection of the join
     * @tparam O The return type of the projection, for which type information must be known
-    * @return A fully co-grouped data set of Os
+    * @return A data set of Os
     */
-  def projecting[O: TypeInformation: ClassTag](fun: (Stream[L], Stream[R]) => O): DataSet[O] =
-    dataset {
-      (left, right) =>
-        fun(left.toStream, right.toStream)
-    }
+  def projecting[O: TypeInformation: ClassTag](fun: (L, R) => O): DataSet[O] =
+    ds(fun)
 
 }
