+++ flink-scala/src/main/scala/org/apache/flink/api/scala/extensions/impl/acceptPartialFunctions/OnJoinFunctionAssigner.scala
@@ -15,24 +15,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.api.scala.extensions.acceptPartialFunctions
+package org.apache.flink.api.scala.extensions.impl.acceptPartialFunctions
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.api.scala.{DataSet, CrossDataSet}
+import org.apache.flink.api.scala.{DataSet, JoinFunctionAssigner}
 
 import scala.reflect.ClassTag
 
-class OnCrossDataSet[L: TypeInformation, R: TypeInformation](ds: CrossDataSet[L, R]) {
+class OnJoinFunctionAssigner[L: TypeInformation, R: TypeInformation](dataset: JoinFunctionAssigner[L, R]) {
 
   /**
-    * Starting from a cross data set, uses the function `fun` to project elements from
-    * both the input data sets in the resulting data set
+    * Joins the data sets using the function `fun` to project elements from both in the
+    * resulting data set
     *
     * @param fun The function that defines the projection of the join
     * @tparam O The return type of the projection, for which type information must be known
-    * @return A data set of Os
+    * @return A fully joined data set of Os
     */
   def projecting[O: TypeInformation: ClassTag](fun: (L, R) => O): DataSet[O] =
-    ds(fun)
+    dataset(fun)
 
 }
