+++ flink-scala/src/main/scala/org/apache/flink/api/scala/extensions/package.scala
@@ -18,37 +18,68 @@
 package org.apache.flink.api.scala
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.api.scala.extensions.acceptPartialFunctions._
+import org.apache.flink.api.scala.extensions.impl.acceptPartialFunctions._
 
 import scala.reflect.ClassTag
 
+/**
+  * acceptPartialFunctions extends the original DataSet with methods with unique names
+  * that delegate to core higher-order functions (e.g. `map`) so that we can work around
+  * the fact that overloaded methods taking functions as parameters can't accept partial
+  * functions as well. This enables the possibility to directly apply pattern matching
+  * to decompose inputs such as tuples, case classes and collections.
+  *
+  * e.g.
+  * {{{
+  *   object Main {
+  *     import org.apache.flink.api.scala.extensions._
+  *     case class Point(x: Double, y: Double)
+  *     def main(args: Array[String]): Unit = {
+  *       val env = ExecutionEnvironment.getExecutionEnvironment
+  *       val ds = env.fromElements(Point(1, 2), Point(3, 4), Point(5, 6))
+  *       ds.filterWith {
+  *         case Point(x, _) => x > 1
+  *       }.reduceWith {
+  *         case (Point(x1, y1), (Point(x2, y2))) => Point(x1 + y1, x2 + y2)
+  *       }.mapWith {
+  *         case Point(x, y) => (x, y)
+  *       }.flatMapWith {
+  *         case (x, y) => Seq('x' -> x, 'y' -> y)
+  *       }.groupingBy {
+  *         case (id, value) => id
+  *       }
+  *     }
+  *   }
+  * }}}
+  *
+  */
 package object extensions {
 
-  implicit def acceptPartialFunctionsOnDataSet[T: TypeInformation](ds: DataSet[T]): OnDataSet[T] =
+  implicit def acceptPartialFunctions[T: TypeInformation](ds: DataSet[T]): OnDataSet[T] =
     new OnDataSet[T](ds)
 
-  implicit def acceptPartialFunctionsOnJoinFunctionAssigner[L: TypeInformation, R: TypeInformation](
+  implicit def acceptPartialFunctions[L: TypeInformation, R: TypeInformation](
       ds: JoinFunctionAssigner[L, R]): OnJoinFunctionAssigner[L, R] =
     new OnJoinFunctionAssigner[L, R](ds)
 
-  implicit def acceptPartialFunctionsOnCrossDataSet[L: TypeInformation, R: TypeInformation](
+  implicit def acceptPartialFunctions[L: TypeInformation, R: TypeInformation](
       ds: CrossDataSet[L, R]): OnCrossDataSet[L, R] =
     new OnCrossDataSet[L, R](ds)
 
-  implicit def acceptPartialFunctionsOnGroupedDataSet[T: TypeInformation: ClassTag](
+  implicit def acceptPartialFunctions[T: TypeInformation: ClassTag](
       ds: GroupedDataSet[T]):
       OnGroupedDataSet[T] =
     new OnGroupedDataSet[T](ds)
 
-  implicit def acceptPartialFunctionsOnCoGroupDataSet[L: TypeInformation, R: TypeInformation](
+  implicit def acceptPartialFunctions[L: TypeInformation, R: TypeInformation](
       ds: CoGroupDataSet[L, R]): OnCoGroupDataSet[L, R] =
     new OnCoGroupDataSet[L, R](ds)
 
-  implicit def acceptPartialFunctionsOnHalfUnfinishedKeyPairOperation[L: TypeInformation, R: TypeInformation, O: TypeInformation](
+  implicit def acceptPartialFunctions[L: TypeInformation, R: TypeInformation, O: TypeInformation](
       ds: HalfUnfinishedKeyPairOperation[L, R, O]): OnHalfUnfinishedKeyPairOperation[L, R, O] =
     new OnHalfUnfinishedKeyPairOperation[L, R, O](ds)
 
-  implicit def acceptPartialFunctionsOnUnfinishedKeyPairOperation[L: TypeInformation, R: TypeInformation, O: TypeInformation](
+  implicit def acceptPartialFunctions[L: TypeInformation, R: TypeInformation, O: TypeInformation](
       ds: UnfinishedKeyPairOperation[L, R, O]): OnUnfinishedKeyPairOperation[L, R, O] =
     new OnUnfinishedKeyPairOperation[L, R, O](ds)
